extract FeaturedStrategies component
